Clean up spec tests
Why you made the change:

I did this to refocus test coverage on our integration tests as we see
more value in those. I also stripped down the isolation tests a bit
since the cases are now covered in the integration tests.

Beyond that I changed a test that was stubbing nil to test for non call
to a not_to raise_error expectation so that RSpec won't be monkey
patching the nil object.