Add #wrapped_logger method
Why you made the change:

I did this so that users would have a way of retrieving the logger that
they handed into the OptionalLogger::Logger during construction. This is
specifically useful when say a gem that is using OptionalLogger wants to
be a good citizen and pass the original logger onto another gem that may
or may not be using OptionalLogger. See issue #15 for more details.