Add logger management instance support
Why you made the change:

I did this so that the logger management module can be used at either
the class level using extend, or at the instance level using include.
This gives the management interface far more flexibility and allows it
to be used in the majority of situations. This is definitely a breaking
change in terms of the public API so this will result in a major version
bump. But, given our current usage is pretty small I think it is worth
it.