Bump version to v2.0.0
Why you made the change:

I bumped this to v2.0.0 a major version bump because we changed the
logger management interface which is a public interface. Therefore, it
is NOT a backwards compatible change and requires consumers of this
library to alter how they use this gem. On the positive side the logger
interface itself hasn't changed so consumer gems upgrading to this will
likely NOT have to break their interfaces, in turn allowing them to do a
patch/minor bump.